+++ app/src/main/java/com/shub39/rush/data/datastore/OtherPreferencesImpl.kt
@@ -14,7 +14,7 @@
  * You should have received a copy of the GNU General Public License
  * along with this program.  If not, see <https://www.gnu.org/licenses/>.
  */
-package com.shub39.rush.data
+package com.shub39.rush.data.datastore
 
 import android.util.Log
 import androidx.datastore.core.DataStore
