+++ app/src/main/java/com/shub39/rush/data/network/GeniusApi.kt
@@ -18,7 +18,6 @@ package com.shub39.rush.data.network
 
 import com.shub39.rush.BuildConfig
 import com.shub39.rush.data.network.dto.genius.GeniusSearchDto
-import com.shub39.rush.data.safeCall
 import com.shub39.rush.domain.Result
 import io.ktor.client.HttpClient
 import io.ktor.client.engine.okhttp.OkHttp
