+++ app/src/main/java/com/shub39/rush/data/network/LyricsPlusApi.kt
@@ -17,7 +17,6 @@
 package com.shub39.rush.data.network
 
 import com.shub39.rush.BuildConfig
-import com.shub39.rush.data.safeCall
 import com.shub39.rush.domain.Result
 import com.shub39.rush.domain.dataclasses.TTMLParser
 import io.ktor.client.HttpClient
