+++ app/src/main/java/com/shub39/rush/di/RushModules.kt
@@ -23,10 +23,10 @@ import coil3.ImageLoader
 import coil3.disk.DiskCache
 import coil3.request.CachePolicy
 import coil3.request.crossfade
-import com.shub39.rush.data.DatastoreFactory
-import com.shub39.rush.data.LyricsPagePreferencesImpl
-import com.shub39.rush.data.OtherPreferencesImpl
-import com.shub39.rush.data.SharePagePreferencesImpl
+import com.shub39.rush.data.datastore.DatastoreFactory
+import com.shub39.rush.data.datastore.LyricsPagePreferencesImpl
+import com.shub39.rush.data.datastore.OtherPreferencesImpl
+import com.shub39.rush.data.datastore.SharePagePreferencesImpl
 import com.shub39.rush.data.database.DatabaseFactory
 import com.shub39.rush.data.database.SongDao
 import com.shub39.rush.data.database.SongDatabase
