+++ app/src/main/java/com/shub39/rush/presentation/ListItemExt.kt
@@ -56,4 +56,4 @@ fun endItemShape(): Shape =
         bottomStart = endCornerRadius.dp,
     )
 
-fun detachedItemShape(): Shape = RoundedCornerShape(1000.dp)
+fun detachedItemShape(): Shape = RoundedCornerShape(endCornerRadius.dp)
