+++ app/src/main/java/com/shub39/rush/presentation/component/ChangelogSheet.kt
@@ -14,7 +14,7 @@
  * You should have received a copy of the GNU General Public License
  * along with this program.  If not, see <https://www.gnu.org/licenses/>.
  */
-package com.shub39.rush.presentation
+package com.shub39.rush.presentation.component
 
 import androidx.compose.foundation.background
 import androidx.compose.foundation.layout.Arrangement
@@ -46,7 +46,13 @@ import androidx.compose.ui.res.stringResource
 import androidx.compose.ui.unit.dp
 import com.shub39.rush.R
 import com.shub39.rush.app.VersionEntry
-import com.shub39.rush.presentation.components.RushBottomSheet
+import com.shub39.rush.presentation.detachedItemShape
+import com.shub39.rush.presentation.endItemShape
+import com.shub39.rush.presentation.leadingItemShape
+import com.shub39.rush.presentation.listItemColors
+import com.shub39.rush.presentation.middleItemShape
+import com.shub39.rush.presentation.theme.flexFontEmphasis
+import com.shub39.rush.presentation.theme.flexFontRounded
 
 @OptIn(ExperimentalMaterial3Api::class, ExperimentalMaterial3ExpressiveApi::class)
 @Composable
