+++ app/src/main/java/com/shub39/rush/presentation/lyrics/LyricsGraph.kt
@@ -37,7 +37,7 @@ import com.google.accompanist.permissions.isGranted
 import com.google.accompanist.permissions.rememberPermissionState
 import com.shub39.rush.domain.dataclasses.Theme
 import com.shub39.rush.navigation.horizontalTransitionMetadata
-import com.shub39.rush.presentation.components.RushTheme
+import com.shub39.rush.presentation.theme.RushTheme
 import com.shub39.rush.presentation.lyrics.section.LyricsCustomisationsPage
 import com.shub39.rush.presentation.lyrics.section.LyricsPage
 import com.shub39.rush.presentation.updateSystemBars
