+++ app/src/main/java/com/shub39/rush/presentation/lyrics/Util.kt
@@ -41,8 +41,8 @@ import com.shub39.rush.domain.dataclasses.WaveColors
 import com.shub39.rush.domain.enums.CardColors
 import com.shub39.rush.domain.enums.LyricsAlignment
 import com.shub39.rush.domain.enums.LyricsBackground
-import com.shub39.rush.presentation.components.ArtFromUrl
-import com.shub39.rush.presentation.components.HypnoticVisualizer
+import com.shub39.rush.presentation.component.ArtFromUrl
+import com.shub39.rush.presentation.component.HypnoticVisualizer
 import com.shub39.rush.presentation.generateGradientColors
 import com.shub39.rush.presentation.lyrics.component.CurveVisualizer
 import com.shub39.rush.presentation.lyrics.component.GradientVisualizer
