+++ app/src/main/java/com/shub39/rush/presentation/lyrics/section/LyricsCustomisationPage.kt
@@ -64,10 +64,10 @@ import com.shub39.rush.domain.dataclasses.Theme
 import com.shub39.rush.domain.enums.AppTheme
 import com.shub39.rush.domain.enums.CardColors
 import com.shub39.rush.domain.enums.LyricsBackground
-import com.shub39.rush.presentation.components.ColorPickerDialog
-import com.shub39.rush.presentation.components.RushDialog
-import com.shub39.rush.presentation.components.RushTheme
-import com.shub39.rush.presentation.flexFontEmphasis
+import com.shub39.rush.presentation.component.ColorPickerDialog
+import com.shub39.rush.presentation.component.RushDialog
+import com.shub39.rush.presentation.theme.RushTheme
+import com.shub39.rush.presentation.theme.flexFontEmphasis
 import com.shub39.rush.presentation.lyrics.LyricsPageAction
 import com.shub39.rush.presentation.lyrics.LyricsPageState
 import com.shub39.rush.presentation.lyrics.component.customisation.LyricsCustomisationPreview
