+++ app/src/main/java/com/shub39/rush/presentation/lyrics/section/LyricsPage.kt
@@ -76,14 +76,14 @@ import com.shub39.rush.domain.enums.LyricsAlignment
 import com.shub39.rush.domain.enums.LyricsBackground
 import com.shub39.rush.presentation.KeepScreenOn
 import com.shub39.rush.presentation.audioDependentBackgrounds
-import com.shub39.rush.presentation.components.ArtFromUrl
-import com.shub39.rush.presentation.components.Empty
-import com.shub39.rush.presentation.components.PageFill
-import com.shub39.rush.presentation.components.RushTheme
+import com.shub39.rush.presentation.component.ArtFromUrl
+import com.shub39.rush.presentation.component.Empty
+import com.shub39.rush.presentation.component.PageFill
+import com.shub39.rush.presentation.theme.RushTheme
 import com.shub39.rush.presentation.fadeBottomToTop
 import com.shub39.rush.presentation.fadeTopToBottom
-import com.shub39.rush.presentation.flexFontEmphasis
-import com.shub39.rush.presentation.flexFontRounded
+import com.shub39.rush.presentation.theme.flexFontEmphasis
+import com.shub39.rush.presentation.theme.flexFontRounded
 import com.shub39.rush.presentation.glowBackground
 import com.shub39.rush.presentation.lyrics.ApplyLyricsBackground
 import com.shub39.rush.presentation.lyrics.LyricsPageAction
