+++ app/src/main/java/com/shub39/rush/presentation/onboarding/Onboarding.kt
@@ -62,8 +62,8 @@ import androidx.compose.ui.unit.dp
 import com.shub39.rush.R
 import com.shub39.rush.domain.dataclasses.Theme
 import com.shub39.rush.domain.enums.AppTheme
-import com.shub39.rush.presentation.components.RushTheme
-import com.shub39.rush.presentation.flexFontEmphasis
+import com.shub39.rush.presentation.theme.RushTheme
+import com.shub39.rush.presentation.theme.flexFontEmphasis
 import com.shub39.rush.presentation.onboarding.component.AnimatedAppIcon
 import kotlinx.coroutines.delay
 import kotlinx.coroutines.launch
