+++ app/src/main/java/com/shub39/rush/presentation/saved/SavedPage.kt
@@ -74,14 +74,14 @@ import com.shub39.rush.domain.dataclasses.Theme
 import com.shub39.rush.domain.enums.AppTheme
 import com.shub39.rush.domain.enums.PaletteStyle
 import com.shub39.rush.domain.enums.SortOrder
-import com.shub39.rush.presentation.components.ArtFromUrl
-import com.shub39.rush.presentation.components.Empty
-import com.shub39.rush.presentation.components.RushTheme
-import com.shub39.rush.presentation.flexFontEmphasis
-import com.shub39.rush.presentation.flexFontRounded
+import com.shub39.rush.presentation.component.ArtFromUrl
+import com.shub39.rush.presentation.component.Empty
+import com.shub39.rush.presentation.theme.RushTheme
+import com.shub39.rush.presentation.theme.flexFontEmphasis
+import com.shub39.rush.presentation.theme.flexFontRounded
 import com.shub39.rush.presentation.saved.component.SavedPageActions
 import com.shub39.rush.presentation.saved.component.SongCard
-import com.shub39.rush.presentation.simpleVerticalScrollbar
+import com.shub39.rush.presentation.component.simpleVerticalScrollbar
 import com.shub39.rush.presentation.toStringRes
 
 @OptIn(ExperimentalMaterial3Api::class, ExperimentalMaterial3ExpressiveApi::class)
