+++ app/src/main/java/com/shub39/rush/presentation/searchsheet/SearchSheet.kt
@@ -66,7 +66,7 @@ import com.shub39.rush.R
 import com.shub39.rush.domain.SourceError
 import com.shub39.rush.domain.dataclasses.Theme
 import com.shub39.rush.domain.enums.AppTheme
-import com.shub39.rush.presentation.components.RushTheme
+import com.shub39.rush.presentation.theme.RushTheme
 import com.shub39.rush.presentation.errorStringRes
 import com.shub39.rush.presentation.lyrics.component.ErrorCard
 import com.shub39.rush.presentation.searchsheet.component.SearchResultCard
