+++ app/src/main/java/com/shub39/rush/presentation/searchsheet/component/SearchResultCard.kt
@@ -33,7 +33,7 @@ import androidx.compose.ui.text.style.TextOverflow
 import androidx.compose.ui.unit.dp
 import com.shub39.rush.R
 import com.shub39.rush.domain.dataclasses.SearchResult
-import com.shub39.rush.presentation.components.ArtFromUrl
+import com.shub39.rush.presentation.component.ArtFromUrl
 
 @Composable
 fun SearchResultCard(result: SearchResult, onClick: () -> Unit, downloaded: Boolean = false) {
