+++ app/src/main/java/com/shub39/rush/presentation/setting/component/AboutApp.kt
@@ -44,9 +44,9 @@ import androidx.compose.ui.unit.dp
 import com.shub39.rush.BuildConfig
 import com.shub39.rush.R
 import com.shub39.rush.domain.dataclasses.Theme
-import com.shub39.rush.presentation.components.RushTheme
+import com.shub39.rush.presentation.theme.RushTheme
 import com.shub39.rush.presentation.endItemShape
-import com.shub39.rush.presentation.flexFontRounded
+import com.shub39.rush.presentation.theme.flexFontRounded
 import com.shub39.rush.presentation.leadingItemShape
 import com.shub39.rush.presentation.middleItemShape
