+++ app/src/main/java/com/shub39/rush/presentation/setting/section/BackupPage.kt
@@ -60,10 +60,10 @@ import com.shub39.rush.domain.backup.ExportState
 import com.shub39.rush.domain.backup.RestoreState
 import com.shub39.rush.domain.dataclasses.Theme
 import com.shub39.rush.domain.enums.AppTheme
-import com.shub39.rush.presentation.components.PageFill
-import com.shub39.rush.presentation.components.RushTheme
+import com.shub39.rush.presentation.component.PageFill
+import com.shub39.rush.presentation.theme.RushTheme
 import com.shub39.rush.presentation.endItemShape
-import com.shub39.rush.presentation.flexFontEmphasis
+import com.shub39.rush.presentation.theme.flexFontEmphasis
 import com.shub39.rush.presentation.leadingItemShape
 import com.shub39.rush.presentation.listItemColors
 import com.shub39.rush.presentation.setting.SettingsPageAction
