+++ app/src/main/java/com/shub39/rush/presentation/theme/ProvideTypography.kt
@@ -14,7 +14,7 @@
  * You should have received a copy of the GNU General Public License
  * along with this program.  If not, see <https://www.gnu.org/licenses/>.
  */
-package com.shub39.rush.presentation
+package com.shub39.rush.presentation.theme
 
 import androidx.compose.foundation.layout.Column
 import androidx.compose.material3.Text
